+++ src/script/utils/interfaces.components.ts
